[标准化libc] fix CI warning

+++ components/libc/compilers/common/none-gcc/sys/unistd.h
@@ -18,8 +18,8 @@
 #define STDOUT_FILENO   1       /* standard output file descriptor */
 #define STDERR_FILENO   2       /* standard error file descriptor */
 
-int read(int fd, void *buf, size_t len);
-int write(int fd, const void *buf, size_t len);
+ssize_t read(int fd, void *buf, size_t len);
+ssize_t write(int fd, const void *buf, size_t len);
 off_t lseek(int fd, off_t offset, int whence);
 int fsync(int fildes);
 int unlink(const char *pathname);
